backend/demo_requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.7.0
python-dotenv==1.0.0
groq